# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PhpShop",
        "description": "API documentation for your Laravel project",
        "contact": {
            "email": "ab.xach@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/reviews/{articleId}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get reviews",
                "description": "Get reviews for article",
                "operationId": "62b508d104951bc234f593b69c4bae71",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Article id",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews for article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reviews retrieved successfully"
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "properties": {
                                                "reviews": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "example": "111"
                                                            },
                                                            "rating": {
                                                                "type": "integer",
                                                                "example": "4"
                                                            },
                                                            "comment": {
                                                                "type": "string",
                                                                "example": "nice article"
                                                            },
                                                            "user": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string",
                                                                        "example": "111"
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Marcus"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/reviews": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Post review",
                "description": "Post review",
                "operationId": "9fe745d363761b681b0e4f1d129e2456",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "article_id",
                                    "rating",
                                    "comment"
                                ],
                                "properties": {
                                    "article_id": {
                                        "type": "integer",
                                        "example": "63"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "nice"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "111"
                                                        },
                                                        "rating": {
                                                            "type": "integer",
                                                            "example": "3"
                                                        },
                                                        "comment": {
                                                            "type": "string",
                                                            "example": "nice"
                                                        },
                                                        "user_id": {
                                                            "type": "string",
                                                            "example": "1"
                                                        },
                                                        "article_id": {
                                                            "type": "number",
                                                            "format": "double",
                                                            "example": "33.33"
                                                        },
                                                        "order_date": {
                                                            "type": "string",
                                                            "example": "2025-03-30 03:18:10"
                                                        },
                                                        "delivery_date": {
                                                            "type": "string",
                                                            "example": "null"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reviews/{reviewId}": {
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update review",
                "description": "Update review",
                "operationId": "6e69f8b89f818600d713ddcb5d65997f",
                "parameters": [
                    {
                        "name": "reviewId",
                        "in": "path",
                        "description": "Review id",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 6
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "comment"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "nice"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order updated successfully."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "111"
                                                        },
                                                        "rating": {
                                                            "type": "integer",
                                                            "example": "3"
                                                        },
                                                        "comment": {
                                                            "type": "string",
                                                            "example": "nice"
                                                        },
                                                        "user_id": {
                                                            "type": "string",
                                                            "example": "1"
                                                        },
                                                        "article_id": {
                                                            "type": "number",
                                                            "format": "double",
                                                            "example": "33.33"
                                                        },
                                                        "order_date": {
                                                            "type": "string",
                                                            "example": "2025-03-30 03:18:10"
                                                        },
                                                        "delivery_date": {
                                                            "type": "string",
                                                            "example": "null"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete review",
                "description": "Delete review",
                "operationId": "f42c5337c3830603204e8b439ac3b123",
                "parameters": [
                    {
                        "name": "reviewId",
                        "in": "path",
                        "description": "Review id",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order updated successfully."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: 'Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}